requirements.txt
torch~=2.5.1
torchvision~=0.20.1
torchaudio
accelerate~=1.2.1
evaluate
datasets~=3.2.0
transformers~=4.47.1
tensorboard
numpy~=1.26.4
scipy
diffusers==0.30.2
wandb~=0.19.1
bitsandbytes
nvidia-ml-py3
sagemaker-ssh-helper
urllib3
requests-toolbelt
torchtools
webdataset~=0.2.100
peft
SentencePiece
opencv-python~=4.10.0.84

pydantic~=2.10.4
pillow~=11.0.0
tqdm~=4.67.1
AnyPathLib~=0.2.0
pandas~=2.2.3
